requirements.txt
flask
python-dotenv
langchain
langchain-openai
openai
pinecone-client
PyPDF2
python-docx
tqdm
